Add form to edit page
